requirements.txt
pandas
numpy
scikit-learn
requests
python-dotenv
seaborn
matplotlib
yfinance
ta
features
hopsworks
polygon-api-client